--- beats-rust/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "beats"
-version = "0.1.0"
-edition = "2021"
-
-build = "build.rs"
-
-[lib]
-crate-type = ["staticlib"]
