--- beats-rust/Makefile.toml
@@ -1,20 +1,0 @@
-# use: cargo install cargo-make / cargo make beats-linux / cargo make beats-windows
-[tasks.build-release]
-command = "cargo"
-args = ["build", "--release"]
-
-[tasks.move-lib-windows]
-script = [
-    "move target\\release\\beats.lib .\\beats.lib"
-]
-
-[tasks.move-lib-linux]
-script = [
-    "mv target/release/beats.lib ./beats.lib"
-]
-
-[tasks.beats-windows]
-dependencies = ["build-release", "move-lib-windows"]
-
-[tasks.beats-linux]
-dependencies = ["build-release", "move-lib-linux"]
