+++ rust/.gitignore
@@ -1,0 +1,4 @@
+/target
+/Cargo.lock
+*.lib
+*.a
